OBJ-70 Moving upload controller into upload folder and adding JSDoc
