+++ src/controllers/upload/upload.responder.strategy.factory.ts
