+++ src/controllers/upload/upload.responder.strategy.ts
